Allow picking the date in your locale and timezone
This is using the Tempus Dominus DatePicker and a custom widget.
The DatePicker only sends the localized date which is not very useful.
The workaround is, that a hidden Field will be updated simultaneously.
The custom transformer then just ignores the value that the datepicker
sends and only uses the ISO date.