Fix the login with "Frame Embed Redirect" configured for the domain
Heroku does not allow adding domains unless the account contains billing
information (Credit Card).
So the only way to have the URL not change from fronds.tv to
fronds-scheduler.herokuapp.com is to configure the domain as a "Frame
embed redirect" where the whole page is embedded as a frame.
That messes with the login though.
The workaround is to open the herokuapp domain in a new tab, when the
login link is clicked.

This should go away if the application runs on a proper server