Remove time restriction of the calendar view
Restricting the time to only show stuff between 10:00 and 0:00 is bad
because the event times are adjusted to the timezone of the visitor.
That can lead to events going unnoticed.
A better solution is to at least scroll to a reasonable time (10:00 am)
but still allow the user to scroll back.